Converting the dimensions of the DataContainer to use 64 bit signed integers. This required LOTS of changes to almost every file. Also, HDF5 files that were written prior to this will need to be rewritten due to changes in the size of the integers written; 4 byte vs 8 byte integers for dimensions and number of points

+++ Source/DREAM3DLib/HDF5/H5VoxelWriter.cpp
@@ -57,7 +57,7 @@ H5VoxelWriter::~H5VoxelWriter()
 int H5VoxelWriter::writeEulerData(float* e1,
                                   float* e2,
                                   float* e3,
-                                  size_t totalPoints,
+                                  int64_t totalPoints,
                                   bool appendFile)
 {
   int err = -1;
@@ -95,7 +95,7 @@ int H5VoxelWriter::writeEulerData(float* e1,
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-int H5VoxelWriter::writeGrainIds(int* grain_indicies, size_t totalPoints, bool appendFile)
+int H5VoxelWriter::writeGrainIds(int* grain_indicies, int64_t totalPoints, bool appendFile)
 {
   int err = -1;
   AIM_H5VtkDataWriter::Pointer h5writer = AIM_H5VtkDataWriter::New();
@@ -125,7 +125,7 @@ int H5VoxelWriter::writeGrainIds(int* grain_indicies, size_t totalPoints, bool a
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-int H5VoxelWriter::writePhaseIds(int* phases, size_t totalPoints, bool appendFile)
+int H5VoxelWriter::writePhaseIds(int* phases, int64_t totalPoints, bool appendFile)
 {
   int err = -1;
   AIM_H5VtkDataWriter::Pointer h5writer = AIM_H5VtkDataWriter::New();
@@ -156,7 +156,7 @@ int H5VoxelWriter::writePhaseIds(int* phases, size_t totalPoints, bool appendFil
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-int H5VoxelWriter::writeStructuredPoints(int volDims[3], float spacing[3], float origin[3], bool appendFile)
+int H5VoxelWriter::writeStructuredPoints(int64_t volDims[3], float spacing[3], float origin[3], bool appendFile)
 {
   int err = -1;
   AIM_H5VtkDataWriter::Pointer h5writer = AIM_H5VtkDataWriter::New();
